refactor: move code to more intuitive places
